src/test/10/requirements.txt
torch
transformers
safetensors
accelerate
diffusers
scipy
invisible_watermark
xformers
tk
pillow
Flask